Add language support for incomplete filter parameters
We want to add autocomplete behavior for the case where a developer is
typing out a liquid statement but isn't finished yet:

```liquid
{{ product | image_url: █
```

The scenario above works without any extra code but if we were to do:

```liquid
{{ product | image_url: width: 100, h█
```

Our parser would blow up because this isn't valid liquid anymore. It
interprets the `h` as a positional argument which cannot come after a
named argument. To support this we need to override how the grammar for
liquid filters is defined in the Placeholder scenario to support this.
The changes will treat the `h` as a `LiquidVariableOutput` and include
proper hierarchy.